Refactor entertainment API into modular components
