+++ package.json
@@ -1,3 +1,4 @@
 {
+  "type": "module",
   "dependencies": {}
 }
